session manager

+++ app/src/main/java/com/example/asodes/SessionManager.kt
@@ -1,5 +1,5 @@
 package com.example.asodes
 
 object SessionManager {
-    var userId: Int? = null
+    var userId: Long? = null
 }
